+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [tanjun.annotations.with_annotated_args][] will now raise if a slash command-specific type is
   used for a required message command parameter.
 - Converter startup client config logs now only use info level when REST fallback is possible.
+- `allowed_types=[]` is now treated as allowing no channel types (rather than any) when passed to
+  [tanjun.conversion.ToChannel.__init__][]
 
 ### Deprecated
 - `BaseConverter.async_caches`, `BaseConverter.cache_components` and `BaseConverter.intents`.